+++ edk2/MdeModulePkg/Core/Dxe/Image/Image.c
@@ -1202,7 +1202,7 @@ CoreStartImage (
   DEBUG_CODE_BEGIN ();
     if (Image->ExitDataSize != 0 || Image->ExitData != NULL) {
 
-      DEBUG ((DEBUG_LOAD, "StartImage: ExitDataSize %d, ExitData %p", Image->ExitDataSize, Image->ExitData));
+      DEBUG ((DEBUG_LOAD, "StartImage: ExitDataSize %d, ExitData %p", (UINT32)Image->ExitDataSize, Image->ExitData));
       if (Image->ExitData != NULL) {
         DEBUG ((DEBUG_LOAD, " (%hs)", Image->ExitData));
       }
